tree: run TestPrettyData in a regular config
In 9c67973e4f751c353b37a5c33e6889f0c6e43e90 TestPrettyData was split up
into short and long versions, with the long one being scheduled to run
nightly. I'm not sure whether it was actually running nightly at some
point, but right now it doesn't - we don't have any CI targets that
specify `go:build nightly` tag. This commit moves the long version into
the regular test and skips it under duress. On my laptop it takes about
6s (as opposed to 0.06s for the short version), so hopefully this won't
be a bottleneck in the CI.

Release note: None
